+++ api/src/shared/domain/usecases/create-lcms-release.js
@@ -1,5 +1,5 @@
-import { lcms } from '../../../src/shared/infrastructure/lcms.js';
-import { LearningContentCache } from '../../infrastructure/caches/learning-content-cache.js';
+import { LearningContentCache } from '../../../../lib/infrastructure/caches/learning-content-cache.js';
+import { lcms } from '../../infrastructure/lcms.js';
 
 const createLcmsRelease = async function () {
   const learningContent = await lcms.createRelease();
