+++ api/tests/shared/unit/domain/usecases/create-lcms-release_test.js
@@ -1,7 +1,7 @@
-import { createLcmsRelease } from '../../../../lib/domain/usecases/create-lcms-release.js';
-import { learningContentCache } from '../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { lcms } from '../../../../src/shared/infrastructure/lcms.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { learningContentCache } from '../../../../../lib/infrastructure/caches/learning-content-cache.js';
+import { createLcmsRelease } from '../../../../../src/shared/domain/usecases/create-lcms-release.js';
+import { lcms } from '../../../../../src/shared/infrastructure/lcms.js';
+import { expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | UseCase | create-lcms-release', function () {
   it('call the createRelease on the lcms module', async function () {
